github: Refactor extscv/github package
Move all V3Client receivers to v3.go
Move all V4Client receivers to v4.go
Move all shared code into common.go
