--- internal/extsvc/github/user_test.go
@@ -1,66 +1,0 @@
-package github
-
-import (
-	"context"
-	"testing"
-
-	"github.com/google/go-cmp/cmp"
-)
-
-// NOTE: To update VCR for this test, please use the token of "sourcegraph-vcr"
-// for GITHUB_TOKEN, which can be found in 1Password.
-func TestListRepositoryCollaborators(t *testing.T) {
-	tests := []struct {
-		name      string
-		owner     string
-		repo      string
-		wantUsers []*Collaborator
-	}{
-		{
-			name:  "public repo",
-			owner: "sourcegraph-vcr-repos",
-			repo:  "public-org-repo-1",
-			wantUsers: []*Collaborator{
-				{
-					ID:         "MDQ6VXNlcjYzMjkwODUx", // sourcegraph-vcr as owner
-					DatabaseID: 63290851,
-				}, {
-					ID:         "MDQ6VXNlcjY2NDY0Nzcz", // sourcegraph-vcr-amy as organization member
-					DatabaseID: 66464773,
-				},
-			},
-		},
-		{
-			name:  "private repo",
-			owner: "sourcegraph-vcr-repos",
-			repo:  "private-org-repo-1",
-			wantUsers: []*Collaborator{
-				{
-					ID:         "MDQ6VXNlcjYzMjkwODUx", // sourcegraph-vcr as owner
-					DatabaseID: 63290851,
-				}, {
-					ID:         "MDQ6VXNlcjY2NDY0Nzcz", // sourcegraph-vcr-amy as organization member
-					DatabaseID: 66464773,
-				}, {
-					ID:         "MDQ6VXNlcjY2NDY0OTI2", // sourcegraph-vcr-bob as outside collaborator
-					DatabaseID: 66464926,
-				},
-			},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			client, save := newV3TestClient(t, "ListRepositoryCollaborators_"+test.name)
-			defer save()
-
-			users, _, err := client.ListRepositoryCollaborators(context.Background(), test.owner, test.repo, 1)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			if diff := cmp.Diff(test.wantUsers, users); diff != "" {
-				t.Fatalf("Users mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
