+++ internal/extsvc/github/v4.go
@@ -11,9 +11,11 @@ import (
 	"strings"
 	"time"
 
+	"github.com/Masterminds/semver"
 	"github.com/graphql-go/graphql/language/ast"
 	"github.com/graphql-go/graphql/language/parser"
 	"github.com/graphql-go/graphql/language/visitor"
+	"github.com/inconshreveable/log15"
 	"github.com/pkg/errors"
 	"golang.org/x/time/rate"
 
@@ -298,4 +300,170 @@ func unmarshal(data []byte, v interface{}) error {
 		return errors.Wrapf(err, "graphql: cannot unmarshal at offset %d: before %q; after %q", e.Offset, string(data[a:e.Offset]), string(data[e.Offset:b]))
 	}
 	return err
+}
+
+// determineGitHubVersion returns a *semver.Version for the targetted GitHub instance by this client. When an
+// error occurs, we print a warning to the logs but don't fail and return the allMatchingSemver.
+func (c *V4Client) determineGitHubVersion(ctx context.Context) *semver.Version {
+	url := normalizeURL(c.apiURL.String())
+	globalVersionCache.mu.Lock()
+	defer globalVersionCache.mu.Unlock()
+
+	if globalVersionCache.lastReset.IsZero() || time.Now().After(globalVersionCache.lastReset.Add(versionCacheResetTime)) {
+		// Clear cache and set last expiry to now.
+		globalVersionCache.lastReset = time.Now()
+		globalVersionCache.versions = make(map[string]*semver.Version)
+	}
+	if version, ok := globalVersionCache.versions[url]; ok {
+		return version
+	}
+	version := c.fetchGitHubVersion(ctx)
+	globalVersionCache.versions[url] = version
+	return version
+}
+
+func (c *V4Client) fetchGitHubVersion(ctx context.Context) *semver.Version {
+	if c.githubDotCom {
+		return allMatchingSemver
+	}
+
+	var resp struct {
+		InstalledVersion string `json:"installed_version"`
+	}
+	req, err := http.NewRequest("GET", "/meta", nil)
+	if err != nil {
+		log15.Warn("Failed to fetch GitHub enterprise version", "build request", "apiURL", c.apiURL, "err", err)
+		return allMatchingSemver
+	}
+	if _, err = doRequest(ctx, c.apiURL, c.auth, c.rateLimitMonitor, c.httpClient, req, &resp); err != nil {
+		log15.Warn("Failed to fetch GitHub enterprise version", "doRequest", "apiURL", c.apiURL, "err", err)
+		return allMatchingSemver
+	}
+	version, err := semver.NewVersion(resp.InstalledVersion)
+	if err == nil {
+		return version
+	}
+	log15.Warn("Failed to fetch GitHub enterprise version", "parse version", "apiURL", c.apiURL, "resp.InstalledVersion", resp.InstalledVersion, "err", err)
+	return allMatchingSemver
+}
+
+func (c *V4Client) GetAuthenticatedUser(ctx context.Context) (*Actor, error) {
+	var result struct {
+		Viewer Actor `json:"viewer"`
+	}
+	err := c.requestGraphQL(ctx, `query GetAuthenticatedUser {
+    viewer {
+        login
+        avatarUrl
+        url
+    }
+}`, nil, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result.Viewer, nil
+}
+
+// GetReposByNameWithOwner fetches the specified repositories (namesWithOwners)
+// from the GitHub GraphQL API and returns a slice of repositories.
+// If a repository is not found, it will return an error.
+//
+// The maximum number of repositories to be fetched is 30. If more
+// namesWithOwners are given, the method returns an error. 30 is not a official
+// limit of the API, but based on the observation that the GitHub GraphQL does
+// not return results when more than 37 aliases are specified in a query. 30 is
+// the conservative step back from 37.
+//
+// This method does not cache.
+func (c *V4Client) GetReposByNameWithOwner(ctx context.Context, namesWithOwners ...string) ([]*Repository, error) {
+	if len(namesWithOwners) > 30 {
+		return nil, ErrBatchTooLarge
+	}
+
+	query, err := c.buildGetReposBatchQuery(namesWithOwners)
+	if err != nil {
+		return nil, err
+	}
+
+	var result map[string]*Repository
+	err = c.requestGraphQL(ctx, query, map[string]interface{}{}, &result)
+	if err != nil {
+		if gqlErrs, ok := err.(graphqlErrors); ok {
+			for _, err2 := range gqlErrs {
+				if err2.Type == graphqlErrTypeNotFound {
+					log15.Warn("GitHub repository not found", "error", err2)
+					continue
+				}
+				return nil, err
+			}
+		} else {
+			return nil, err
+		}
+	}
+
+	repos := make([]*Repository, 0, len(result))
+	for _, r := range result {
+		if r != nil {
+			repos = append(repos, r)
+		}
+	}
+	return repos, nil
+}
+
+func (c *V4Client) buildGetReposBatchQuery(namesWithOwners []string) (string, error) {
+	var b strings.Builder
+	b.WriteString(c.repositoryFieldsGraphQLFragment())
+	b.WriteString("query {\n")
+
+	for i, pair := range namesWithOwners {
+		owner, name, err := SplitRepositoryNameWithOwner(pair)
+		if err != nil {
+			return "", err
+		}
+		fmt.Fprintf(&b, "repo%d: repository(owner: %q, name: %q) { ", i, owner, name)
+		b.WriteString("... on Repository { ...RepositoryFields } }\n")
+	}
+
+	b.WriteString("}")
+
+	return b.String(), nil
+}
+
+// repositoryFieldsGraphQLFragment returns a GraphQL fragment that contains the fields needed to populate the
+// Repository struct.
+func (c *V4Client) repositoryFieldsGraphQLFragment() string {
+	if c.githubDotCom {
+		return `
+fragment RepositoryFields on Repository {
+	id
+	databaseId
+	nameWithOwner
+	description
+	url
+	isPrivate
+	isFork
+	isArchived
+	isLocked
+	isDisabled
+	viewerPermission
+}
+	`
+	}
+	// Some fields are not yet available on GitHub Enterprise yet
+	// or are available but too new to expect our customers to have updated:
+	// - viewerPermission
+	return `
+fragment RepositoryFields on Repository {
+	id
+	databaseId
+	nameWithOwner
+	description
+	url
+	isPrivate
+	isFork
+	isArchived
+	isLocked
+	isDisabled
+}
+	`
 }
